fiber_smoother.m
Updated version of the fiber_smoother function. New version adds the option to use a weighted polynomial fitting scheme to improve the accuracy of the smoothed tracts. New version is backwards compatible, so it should not affect the reproducibility of the previously published results in Damon et al., J Biomech, 2021.